+++ junowen/src/state/select.rs
@@ -65,7 +65,7 @@ pub fn on_input_players(
     } else {
         None
     };
-    let (p1, p2) = session.enqueue_input_and_dequeue(input_devices.p1_input().0 as u8, delay)?;
+    let (p1, p2) = session.enqueue_input_and_dequeue(input_devices.p1_input().0 as u16, delay)?;
     let input_devices = th19.input_devices_mut();
     input_devices.set_p1_input(Input(p1 as u32));
     input_devices.set_p2_input(Input(p2 as u32));
@@ -86,9 +86,9 @@ pub fn on_input_menu(session: &mut Session, th19: &mut Th19) -> Result<(), RecvE
     };
     let (p1, _p2) = session.enqueue_input_and_dequeue(
         if session.host() {
-            th19.menu_input().0 as u8
+            th19.menu_input().0 as u16
         } else {
-            Input::NULL as u8
+            Input::NULL
         },
         delay,
     )?;
